Create a file for each component to enhance redability
