+++ react/components/EditorContainer/StoreEditor/Styles/StyleEditor/typography/TypeTokensList.tsx
@@ -3,8 +3,7 @@ import { FormattedMessage } from 'react-intl'
 import { RouteComponentProps } from 'react-router'
 
 import StyleEditorHeader from '../StyleEditorHeader'
-import { EditorPath, IdParam } from '../StyleEditorRouter'
-import startCase from 'lodash.startcase'
+import TypeTokenEntry from './TypeTokensEntry'
 
 interface Props extends RouteComponentProps {
   style: Style
@@ -32,25 +31,6 @@ const TypeTokensList: React.FunctionComponent<Props> = ({ history, style }) => {
         ))}
       </div>
     </>
-  )
-}
-
-interface EntryProps {
-  name: string
-  history: RouteComponentProps['history']
-}
-
-const TypeTokenEntry: React.FunctionComponent<EntryProps> = ({
-  name,
-  history,
-}) => {
-  return (
-    <div
-      className="pointer flex justify-between items-center pv6 bb b--muted-4"
-      onClick={() => history.push(EditorPath.typeToken.replace(IdParam, name))}
-    >
-      <span className="f4">{startCase(name)}</span>
-    </div>
   )
 }
